Clean up the ClusterLogForwarder API to follow recommendations
- Move methods to separate files, *_type.go only has type definitions.
- Move all the main CRD spec types (InputSpec, OutputSpec) into one file
  - OutputTypeSpec and children (Kafka, Syslog etc.) still in output_types.go
- Add 'clf' as short name.
- Improved API documentation, fixed some minor API issues (+required vs. +optional)
- Dropped the `insecure` field, not useful and is confusing.
